+++ packages/liveblocks-react/test-d/augmentation.test-d.tsx
@@ -269,6 +269,16 @@ declare global {
 
 // ---------------------------------------------------------
 
+// useStorageStatus()
+{
+  expectType<"not-loaded" | "loading" | "synchronizing" | "synchronized">(
+    classic.useStorageStatus()
+  );
+  expectType<"synchronizing" | "synchronized">(suspense.useStorageStatus());
+}
+
+// ---------------------------------------------------------
+
 // useSelf()
 {
   const me = classic.useSelf();