+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 ## v2.1.0 (not published yet)
 
-### `@liveblocks/core`
+### `@liveblocks/client`
 
 - Various internal refactorings
+
+### `@liveblocks/react`
+
+- Add new hook `useStorageStatus()`, which returns the current storage status of
+  the room, and will re-render your component whenever it changes. This can used
+  to build "Saving..." UIs.
 
 ## v2.0.5